chore: split turn result modal
